+++ juniper_graphql_ws/src/graphql_ws/mod.rs
@@ -40,7 +40,7 @@ pub use self::{
 
 struct ExecutionParams<S: Schema> {
     start_payload: StartPayload<S::ScalarValue>,
-    config: Arc<ConnectionConfig<S::Context>>,
+    config: ConnectionConfig<S::Context>,
     schema: S,
 }
 
@@ -61,7 +61,7 @@ enum ConnectionState<S: Schema, I: Init<S::ScalarValue, S::Context>> {
     PreInit { init: I, schema: S },
     /// Active is the state after a ConnectionInit message has been accepted.
     Active {
-        config: Arc<ConnectionConfig<S::Context>>,
+        config: ConnectionConfig<S::Context>,
         stoppers: HashMap<String, oneshot::Sender<()>>,
         schema: S,
     },
@@ -112,7 +112,7 @@ impl<S: Schema, I: Init<S::ScalarValue, S::Context>> ConnectionState<S, I> {
 
                         (
                             Self::Active {
-                                config: Arc::new(config),
+                                config,
                                 stoppers: HashMap::new(),
                                 schema,
                             },
@@ -430,8 +430,7 @@ where
 
 impl<S, I, T> Sink<T> for Connection<S, I>
 where
-    T: TryInto<ClientMessage<S::ScalarValue>>,
-    T::Error: Error,
+    T: TryInto<ClientMessage<S::ScalarValue>, Error: Error>,
     S: Schema,
     I: Init<S::ScalarValue, S::Context> + Send,
 {
@@ -546,6 +545,7 @@ mod test {
 
     use super::{Connection, ConnectionConfig, ConnectionErrorPayload, DataPayload, StartPayload};
 
+    #[derive(Clone, Copy)]
     struct Context(i32);
 
     impl juniper::Context for Context {}